+++ frontend/src/api/routes/robots/rpaParameterRetrieval.js
@@ -31,4 +31,26 @@ const updateManyParameters = async (parameterObjectsList) => {
   return response;
 };
 
-export { getAllParametersForRobot, updateManyParameters };
+/**
+ * @description Sends a callout to the backend to delete parameters for the given activities
+ * @param {String} robotId Id of the robot that is being used
+ * @param {String} unusedActivityListString Stringified List of Activity Ids
+ */
+const deleteParametersForActivities = (robotId, activityIdList) => {
+  const requestStringParameters = `/robots/parameters/${robotId}`;
+  fetch(requestStringParameters, {
+    body: JSON.stringify({ activityIdList }),
+    method: 'DELETE',
+    headers: {
+      'Content-Type': 'application/json;charset=utf-8',
+    },
+  }).catch((err) => {
+    console.error(err);
+  });
+};
+
+export {
+  getAllParametersForRobot,
+  updateManyParameters,
+  deleteParametersForActivities,
+};
