+++ frontend/src/components/pages/RobotModeler/RobotModeler.jsx
@@ -3,10 +3,10 @@ import { Layout } from 'antd';
 import BpmnModeler from './BpmnModeler/BpmnModeler';
 import HeaderNavbar from '../../multiPageComponents/HeaderNavbar/HeaderNavbar';
 import ModelerSidebar from './ModelerSidebar/ModelerSidebar';
-import { getSsot } from '../../../api/ssotRetrieval';
-import { getAllParametersForRobot } from '../../../api/parameterRetrieval';
-import { getAllAttributes } from '../../../api/attributeRetrieval';
-import { getAllRpaFunctionalities } from '../../../api/applicationAndTaskRetrieval';
+import { getSsot } from '../../../api/routes/robots/robotsRetrieval';
+import { getAllParametersForRobot } from '../../../api/routes/robots/rpaParameterRetrieval';
+import { getAllAttributes } from '../../../api/routes/robots/rpaAttributeRetrieval';
+import { getAllRpaFunctionalities } from '../../../api/routes/functionalities/functionalities';
 import { setRobotId } from '../../../utils/sessionStorage/localSsotController/ssot';
 import {
   initAvailableApplicationsSessionStorage,
